+++ graphquery/core/models/lpg_mmap/transaction.h
@@ -27,15 +27,15 @@ namespace graphquery::database::storage
 
         struct SHeaderBlock
         {
-            uint64_t transaction_c           = {};
-            uint64_t transactions_start_addr = {};
-            uint64_t eof_addr                = {};
+            uint32_t transaction_c           = {};
+            uint32_t transactions_start_addr = {};
+            uint32_t eof_addr                = {};
         };
 
         struct SVertexCommit
         {
             char label[CFG_LPG_LABEL_LENGTH] = {""};
-            uint64_t optional_id             = {0};
+            uint32_t optional_id             = {0};
             uint16_t property_c              = {0};
             uint8_t remove                   = {0};
         } __attribute__((packed));
@@ -67,7 +67,7 @@ namespace graphquery::database::storage
         void commit_rm_vertex(const ILPGModel::SNodeID & src) noexcept;
         void commit_rm_edge(const ILPGModel::SNodeID & src, const ILPGModel::SNodeID & dst, std::string_view label = "") noexcept;
 
-        void commit_vertex(std::string_view label, const std::vector<ILPGModel::SProperty_t> & props, uint64_t optional_id = -1) noexcept;
+        void commit_vertex(std::string_view label, const std::vector<ILPGModel::SProperty_t> & props, uint32_t optional_id = -1) noexcept;
         void commit_edge(const ILPGModel::SNodeID & src, const ILPGModel::SNodeID & dst, std::string_view label, const std::vector<ILPGModel::SProperty_t> & props) noexcept;
 
       private:
@@ -79,7 +79,7 @@ namespace graphquery::database::storage
         void define_transaction_header();
 
         template<typename T>
-        inline SRef_t<T> read_transaction(uint64_t seek);
+        inline SRef_t<T> read_transaction(uint32_t seek);
         inline SRef_t<SHeaderBlock> read_transaction_header();
 
         void process_vertex_transaction(SRef_t<SVertexTransaction> &, const std::vector<ILPGModel::SProperty_t> & props) const noexcept;
@@ -89,7 +89,7 @@ namespace graphquery::database::storage
         CDiskDriver m_transaction_file;
         static constexpr int64_t INITIAL_TRANSACTION_FILE_SIZE  = KB(4);
         static constexpr const char * TRANSACTION_FILE_NAME     = "transactions";
-        static constexpr uint64_t TRANSACTION_HEADER_START_ADDR = 0x00000000;
-        static constexpr uint64_t TRANSACTIONS_START_ADDR       = sizeof(SHeaderBlock);
+        static constexpr uint32_t TRANSACTION_HEADER_START_ADDR = 0x00000000;
+        static constexpr uint32_t TRANSACTIONS_START_ADDR       = sizeof(SHeaderBlock);
     };
 } // namespace graphquery::database::storage
